+++ cmd/orq/main.go
@@ -36,5 +36,9 @@ func traceAPI() commands.TraceAPI {
 			_, decoded, err := generated.OpenapiTracesListSpans(traceID, params)
 			return decoded, err
 		},
+		GetResponse: func(responseID string, params *viper.Viper) (map[string]any, error) {
+			_, decoded, err := generated.OpenapiRetrieveResponse(responseID, params)
+			return decoded, err
+		},
 	}
 }